Drop `follow-file-renames` (implemented by GitHub)
